Implement PUT /eholdings/configuration endpoint
Jira: MODKBEKBJ-41
Implemented put method the same way it is implemented in ruby module:
1)Check that new configuration is valid by calling RM API
2)Get old configurations from OKAPI
3)Delete old configurations by id
4)Post new configurations to OKAPI

Copied RMAPIService from mod-codex-ekb module,
there is probably a better way to reuse this class

There is a small problem with aspectj: if EholdingsConfigurationImpl directly calls
any methods of model classes (classes that are generated from raml) then build
fails with compilation error on aspectj-maven-plugin step, it seems that
aspectj runs before model is generated and that causes errors.
Moved all calls to model classes to RMAPIConfigurationConverter and ErrorUtil.

Signed-off-by: Andrii Paias <Andrii_Paias@epam.com>