Implement GET /eholdings/providers endpoint
Jira: MODKBEKBJ-18
Implemented get providers method.

Mapped vendors returned from RM API to Providers response the same way
as it is mapped in ruby module.

Renamed ConfigurationUnprocessableError to JsonapiError because the
structure of errors will be the same for all endpoints.

Used @Validate in EholdingsProvidersImpl to validate constraints on page
and count parameters. Added ValidationErrorHandlerAspect to catch validation
errors and wrap them in JsonapiError, otherwise error message is returned
as plain text.