Implement POST /eholdings/resources endpoint
Jira: MODKBEKBJ-30
Implemented POST resources endpoint by sending PUT request to RM API with
isSelected attribute set to true.
Implemented validation to check that package is custom, url is valid and
that title is not already added to package.
Searched titles in package by title name to check that title is not added.